requirements.txt
Flask
firebase-admin
python-dotenv
google-generativeai
flask-cors
